feat: Exception handler fist implementation

+++ src/main/java/com/elves/desafio3/services/ClientService.java
@@ -2,6 +2,7 @@ package com.elves.desafio3.services;
 
 import com.elves.desafio3.dto.ClientDTO;
 import com.elves.desafio3.entities.Client;
+import com.elves.desafio3.exceptions.DomainException;
 import com.elves.desafio3.resources.ClientResource;
 import org.springframework.beans.factory.annotation.Autowired;
 
@@ -10,6 +11,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.NoSuchElementException;
 
 
 @Service
@@ -21,6 +23,7 @@ public class ClientService {
 
     @Transactional(readOnly = true)
     public List<ClientDTO> findAll(){
+
         List<Client> list = resource.findAll();
         List<ClientDTO> result= list.stream().map( x -> new ClientDTO(x)).toList();
 
@@ -30,30 +33,61 @@ public class ClientService {
 
     @Transactional(readOnly = true)
     public ClientDTO findById(Long id){
+    try {
         return new ClientDTO(resource.findById(id).get());
+    }catch (NoSuchElementException e)   {
+        throw new DomainException("Id não encontrado");
+    }
+
     }
 
     @Transactional
-    public ClientDTO insert(Client client){
-        return new ClientDTO(client= resource.save(client));
-    }
+    public ClientDTO insert(ClientDTO client){
 
-    @Transactional
-    public ClientDTO update(Client client, Long id){
-
-        Client entity = resource.getReferenceById(id);
-
+        Client entity = new Client();
+        entity.setId(client.getId());
         entity.setName(client.getName());
         entity.setCpf(client.getCpf());
         entity.setBirthDate(client.getBirthDate());
         entity.setIncome(client.getIncome());
         entity.setChildren(client.getChildren());
-
+        
         return new ClientDTO(resource.save(entity));
+    }
+    @Transactional
+    public ClientDTO update(ClientDTO client, Long id){
+
+        try {
+
+            Client entity = resource.getReferenceById(id);
+
+            entity.setName(client.getName());
+            entity.setCpf(client.getCpf());
+            entity.setBirthDate(client.getBirthDate());
+            entity.setIncome(client.getIncome());
+            entity.setChildren(client.getChildren());
+
+            return new ClientDTO(resource.save(entity));
+
+        }catch (NoSuchElementException e){
+            throw new DomainException("Id não encontrado");
+        }
+
+
     }
 
     @Transactional
     public void delete(Long id){
-        resource.deleteById(id);
+
+        try {
+
+            if(resource.existsById(id)) {
+                resource.deleteById(id);
+            }else throw new NoSuchElementException();
+
+        }catch (NoSuchElementException e){
+            throw new DomainException("Id não encontrado");
+        }
     }
+
 }
